[BUGFIX] Avoid creating large images if not possible
If the native width of an original image is less than
the desired width of a srcset definition the creation
of that image must be avoided because else a low
quality image with a higher width than the
original one will be created.

This only causes images with a bigger size but less
quality then the default one.

Resolves #3