[BUGFIX] Set width and height correctly
When the default rendering is active, the image
tag will be rendered with width an height attributes.

These are now set correctly to the value of the
processed image. This is either the width of 360px
if rendering of a small default image is defined.
Or this is the max width defined by typoscript.

From this commit on, images will always be processed
if their width is too high. If there is a maximum width
of 800 px defined in typoscript and the image has a
higher width, the image will be processed.